+++ src/database/sql/Visits.model.sql
@@ -1,12 +1,13 @@
 CREATE TABLE IF NOT EXISTS visits (
     id uuid PRIMARY KEY DEFAULT gen_random_uuid(),
-    patient_id uuid REFERENCES users(id),
-    doctor_id uuid REFERENCES users(id),
+    patient_id uuid REFERENCES users(id) NOT NULL,
+    doctor_id uuid REFERENCES users(id) NOT NULL, -- comes from the frontend when select the doctor's name
     date TIMESTAMP,
     time VARCHAR(10),
-    patient_notes VARCHAR(500) ,
+    patient_notes VARCHAR(500),
     medical_notes VARCHAR(500),
-    status VARCHAR(15) CHECK (status IN ('scheduled', 'completed', 'canceled', 'pending')) DEFAULT 'scheduled',
+    status VARCHAR(15) CHECK (status IN ('scheduled', 'completed', 'cancelled', 'pending', 'active')) DEFAULT 'pending',
+    amount SMALLINT,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );