+++ src/zod/authentication/signup.schema.js
@@ -1,6 +1,6 @@
 import z from "zod";
 
-const loginUserSchema = z.object({
+const signupSchema = z.object({
   email: z.email("Invalid email address"),
   password: z
     .string()
@@ -14,6 +14,11 @@ const loginUserSchema = z.object({
     .refine((value) => /[!@#$%^&*(),.?":{}|<>]/.test(value), {
       message: "Password must contain at least one special character",
     }),
+  role: z.enum(["doctor", "patient", "finance"]),
+  name: z
+    .string()
+    .min(3, "Name must be at least 3 characters long")
+    .max(20, "Name must be at most 20 characters long"),
 });
 
-export { loginUserSchema };
+export { signupSchema };