+++ src/zod/visits/getAllVisits.schema.js
@@ -1,0 +1,8 @@
+import { z } from "zod";
+
+const getAllVisitsSchema = z.object({
+  id: z.uuid(),
+  role: z.enum(["patient", "doctor", "finance"]),
+});
+
+export { getAllVisitsSchema };